+++ robot_descriptions/talos_description.py
@@ -23,16 +23,20 @@ from os import path as _path
 
 from ._cache import clone_to_cache as _clone_to_cache
 
-PATH: str = _clone_to_cache("talos-data")
+REPOSITORY_PATH: str = _clone_to_cache("talos-data")
 
-MESHES_PATH: str = _path.join(PATH, "meshes")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH)
 
-URDF_PATH: str = _path.join(PATH, "urdf", "talos_full.urdf")
+MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
+
+URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "talos_full.urdf")
 
 # Description-specific paths
 
-URDF_PATH_V2: str = _path.join(PATH, "urdf", "talos_full_v2.urdf")
+URDF_PATH_V2: str = _path.join(PACKAGE_PATH, "urdf", "talos_full_v2.urdf")
 
-URDF_PATH_REDUCED: str = _path.join(PATH, "urdf", "talos_reduced.urdf")
+URDF_PATH_REDUCED: str = _path.join(PACKAGE_PATH, "urdf", "talos_reduced.urdf")
 
-URDF_PATH_REDUCED_V2: str = _path.join(PATH, "urdf", "talos_reduced_v2.urdf")
+URDF_PATH_REDUCED_V2: str = _path.join(
+    PACKAGE_PATH, "urdf", "talos_reduced_v2.urdf"
+)