+++ robot_descriptions/tiago_description.py
@@ -23,16 +23,18 @@ from os import path as _path
 
 from ._cache import clone_to_cache as _clone_to_cache
 
-__working_dir__ = _clone_to_cache("example-robot-data")
+REPOSITORY_PATH: str = _clone_to_cache("example-robot-data")
 
-PATH: str = _path.join(__working_dir__, "robots", "tiago_description")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "robots", "tiago_description")
 
-MESHES_PATH: str = _path.join(PATH, "meshes")
+MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
 
-URDF_PATH: str = _path.join(PATH, "robots", "tiago.urdf")
+URDF_PATH: str = _path.join(PACKAGE_PATH, "robots", "tiago.urdf")
 
 # Description-specific paths
 
-URDF_PATH_DUAL: str = _path.join(PATH, "robots", "tiago_dual.urdf")
+URDF_PATH_DUAL: str = _path.join(PACKAGE_PATH, "robots", "tiago_dual.urdf")
 
-URDF_PATH_NO_HAND: str = _path.join(PATH, "robots", "tiago_no_hand.urdf")
+URDF_PATH_NO_HAND: str = _path.join(
+    PACKAGE_PATH, "robots", "tiago_no_hand.urdf"
+)