+++ robot_descriptions/_repositories.py
@@ -40,6 +40,11 @@ class Repository:
 
 
 REPOSITORIES = {
+    "aliengo_mj_description": Repository(
+        url="https://github.com/rohanpsingh/aliengo_mj_description.git",
+        commit="15051f23ebec30f8b72ed224d435934f6cbc679f",
+        cache_path="rohanpsingh/aliengo_mj_description",
+    ),
     "anymal_b_simple_description": Repository(
         url="https://github.com/ANYbotics/anymal_b_simple_description.git",
         commit="988b5df22b84761bdf08111b1c2ccc883793f456",