+++ robot_descriptions/__init__.py
@@ -19,4 +19,4 @@
 Import open source robot description as Python modules.
 """
 
-__version__ = "0.2.0-pre"
+__version__ = "0.2.0"